Don't link with -lxentoollog for older xen
The -lxentoollog introduced by 77eeff0 is required for newer versions of
xen, including 4.7, but breaks compilation with older versions.

This commit only adds -lxentoollog if we are using a newer xen version,
otherwise it will be omitted. This is done by using an OASIS flag that
is determined by configure.ml, and passed to OASIS in the Makefile.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>